# service/Market/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Order:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      id:
        type: integer
      payment_method:
        type: string
      payment_status:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      seller_id:
        type: integer
      sizes:
        items:
          type: string
        type: array
      status:
        type: string
      stock:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Seller:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      rating:
        type: number
      shop_name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.UpdateOrderStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Marketplace API with products, categories, cart, orders, seller and
    admin management
  title: Market Service API
  version: "1.0"
paths:
  /api/admin/categories:
    post:
      consumes:
      - application/json
      description: Create a new product category (admin only)
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - admin
  /api/admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category (admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update an existing category (admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - admin
  /api/admin/orders:
    get:
      consumes:
      - application/json
      description: Get list of all orders (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - admin
  /api/admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update status of an order (admin only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - admin
  /api/admin/products/{id}/status:
    put:
      consumes:
      - application/json
      description: Update product status (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status data
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update product status
      tags:
      - admin
  /api/admin/sellers:
    get:
      consumes:
      - application/json
      description: Get list of all sellers (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Seller'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all sellers
      tags:
      - admin
  /api/admin/sellers/{id}/status:
    put:
      consumes:
      - application/json
      description: Activate or deactivate a seller (admin only)
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status data
        in: body
        name: request
        required: true
        schema:
          properties:
            is_active:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update seller status
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"